requirements.txt
async-generator==1.10
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
h11==0.13.0
idna==3.3
numpy==1.23.2
opencv-python==4.6.0.66
outcome==1.2.0
Pillow==9.2.0
PySocks==1.7.1
pyTelegramBotAPI==4.7.0
python-dotenv==0.20.0
requests==2.28.1
schedule==1.1.0
selenium==4.4.3
Selenium-Screenshot==2.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
webdriver-manager==3.8.3
wsproto==1.1.0
